+++ tests/testthat/test_get_samples_table.R
@@ -1,0 +1,26 @@
+# capture_requests({
+#   # set a dummy cloudos credentials
+#   my_cloudos <- connect_cloudos(base_url = "http://cohort-browser-766010452.eu-west-1.elb.amazonaws.com",
+#                         auth = Sys.getenv("test_cloudos_apikey"),
+#                         team_id = Sys.getenv("test_cloudos_team_id"))
+#   # set a dummy cohort object
+#   my_cohort <- cb_load_cohort(my_cloudos, cohort_id = "5f6228133097cc7a6504fb76")
+#   # test
+#   sample_table <- cb_get_samples_table(my_cloudos, my_cohort)
+#   expect_s3_class(sample_table, "data.frame")
+# })
+
+with_mock_api({
+  test_that("Get a samples table", {
+    
+    # set a dummy cloudos credentials
+    my_cloudos <- connect_cloudos(base_url = "http://cohort-browser-766010452.eu-west-1.elb.amazonaws.com",
+                          auth = "Bearer token",
+                          team_id = "5f046bf6c132dd15fdd1a525")
+    # set a cohort object
+    my_cohort <- cb_load_cohort(my_cloudos, cohort_id = "5f6228133097cc7a6504fb76")
+    
+    sample_table <- cb_get_samples_table(my_cloudos, my_cohort)
+    expect_s3_class(sample_table, "data.frame")
+  })
+})