+++ NEWS.md
@@ -1,3 +1,3 @@
 # cloudos 0.0.0.9000
 
-* Added a `NEWS.md` file to track changes to the package.
+* First complete set of cb_* functions for Cohort Browser endpoints